substantial refactor/move of files transitioning this repo to a lib only repository.
