+++ cpputest-qpc-lib/include/cmsDummyActiveObject.hpp
