+++ cpputest-qpc-lib/include/cmsTestPublishedEventRecorder.hpp
