+++ cpputest-qpc-lib/include/qevtUniquePtr.hpp
