--- examples/hwLockCtrlService/CMakeLists.txt
@@ -1,10 +1,0 @@
-include_directories(include)
-add_subdirectory(test)
-add_library(hwLockCtrlService include/hwLockCtrlService.h src/hwLockCtrlService.c)
-target_link_libraries(hwLockCtrlService hwLockCtrl)
-target_include_directories(hwLockCtrlService PUBLIC include)
-
-# TODO: setting up QP should probably be a higher level build option
-add_compile_definitions(QACTIVE_CAN_STOP)
-include_directories(${QP_CPP_INCLUDE_DIR})
-include(${EXTERNALS_TOP_DIR}/qpcPosixPortCMakeSupport.txt)
