--- examples/hwLockCtrlService/include/hwLockCtrlServiceExpectedPubSubSigs.h
@@ -1,8 +1,0 @@
-
-HW_LOCK_CTRL_SERVICE_REQUEST_UNLOCKED_SIG,
-HW_LOCK_CTRL_SERVICE_REQUEST_LOCKED_SIG,
-HW_LOCK_CTRL_SERVICE_REQUEST_SELF_TEST_SIG,
-
-HW_LOCK_CTRL_SERVICE_IS_UNLOCKED_SIG,
-HW_LOCK_CTRL_SERVICE_IS_LOCKED_SIG,
-HW_LOCK_CTRL_SERVICE_SELF_TEST_RESULTS_SIG, //uses event type HwLockCtrl::SelfTestEvent
