--- include/bspTicks.h
@@ -1,17 +1,0 @@
-
-#ifndef BSP_TICKS_H
-#define BSP_TICKS_H
-
-#include <stdint.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-static const uint32_t BSP_TICKS_PER_SECOND = 100;
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif   // BSP_TICKS_H
