--- test_support/CMakeLists.txt
@@ -1,2 +1,0 @@
-include_directories(include)
-add_subdirectory(cpputest-qpc-port)
